Add product service

+++ HLL.HLX.BE.Core.Model/Catalog/CrossSellProduct.cs
@@ -11,11 +11,11 @@ namespace HLL.HLX.BE.Core.Model.Catalog
         /// <summary>
         ///     Gets or sets the first product identifier
         /// </summary>
-        public long ProductId1 { get; set; }
+        public int ProductId1 { get; set; }
 
         /// <summary>
         ///     Gets or sets the second product identifier
         /// </summary>
-        public long ProductId2 { get; set; }
+        public int ProductId2 { get; set; }
     }
 }
